feat: resolve conflicts
